[POLY] Added tests cases for double pass typing
* Added GetType() in Form interface (only Pred has a TypeScheme, but
every Form is inherently typed with the default type omicron).
* Updated typing environment to record constants. GetTypeScheme input
arguments are not mandatory anymore, to recover constants TypeScheme.
* Updated GetOutType to take TypeHint as parameter, also for constants
integration.
* Added SaveConstants function because ... parameters should only be at
the end of the parameters list, but it doesn't make any semantic sense,
so this function should now be used to record constants. An interface
function may be created in the future to make generic TypeSchemes (for
constants and non-constants functions or predicates).